Gold/expense-policy-enforcer/backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Email
sendgrid==6.11.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
httpx==0.26.0
